src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/style-insertion-crash-expected.txt

CONSOLE ERROR: line 10: Uncaught TypeError: host.createShadowRoot is not a function
FAIL Unexpected error: Uncaught TypeError: host.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
